refactor: separate models and utilities
